Rename folder to pypsaeur
