+++ src/egon/data/datasets/pypsaeur/Snakefile
